Refactored change password and restaurant info to presentational components
