+++ client/src/components/Presentational/ChangePassword/styles.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import { containerMixin, flexCenterMixin } from '../../global-styles/mixins';
+import { containerMixin, flexCenterMixin } from '../../../global-styles/mixins';
 
 export const Container = styled.div`
   ${containerMixin}
